updated import file location after restructuring folder layout
