+++ client/__test__/CustomerApp/CustomerLogin.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import CustomerLogin from '../src/components/CustomerApp/CustomerLogin';
+import CustomerLogin from '../../src/components/CustomerApp/CustomerLogin';
 
 describe('<CustomerLogin />', () => {
   it('', () => {
