+++ client/__test__/CustomerApp/FindRestaurants.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import FindRestaurants from '../src/components/CustomerApp/FindRestaurants';
+import FindRestaurants from '../../src/components/CustomerApp/FindRestaurants';
 
 describe('<FindRestaurants />', () => {
   it('', () => {
