+++ client/__test__/CustomerApp/Menu.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import Menu from '../src/components/CustomerApp/Menu';
+import Menu from '../../src/components/CustomerApp/Menu';
 
 describe('<Menu />', () => {
   it('', () => {
