+++ client/__test__/CustomerApp/OrderStatus.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import OrderStatus from '../src/components/CustomerApp/OrderStatus';
+import OrderStatus from '../../src/components/CustomerApp/OrderStatus';
 
 describe('<OrderStatus />', () => {
   it('', () => {
