+++ client/__test__/CustomerApp/Restaurant.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import Restaurant from '../src/components/CustomerApp/Restaurant';
+import Restaurant from '../../src/components/CustomerApp/Restaurant';
 
 describe('<Restaurant />', () => {
   it('', () => {
