+++ client/__test__/CustomerApp/RestaurantList.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import RestaurantList from '../src/components/CustomerApp/RestaurantList';
+import RestaurantList from '../../src/components/CustomerApp/RestaurantList';
 
 describe('<RestaurantList />', () => {
   it('', () => {
