+++ client/__test__/RestaurantApp/RestaurantLogin.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../jest.config';
 
-import RestaurantLogin from '../src/components/RestaurantApp/RestaurantLogin';
+import RestaurantLogin from '../../src/components/RestaurantApp/RestaurantLogin';
 
 describe('<RestaurantLogin />', () => {
   it('', () => {
